Fixed ImType being "sci" instead of "focus"
The image type whould default to "sci" because it's takeExposure's
default.
I changed takeExposure's arguments to "ImType='focus'", but because of
this, some lines of code as pretty much useless. (the ones who change
the camera properties to "ImType='focus'")
I didn't remove them because they are harmless. (But YOU, as the person
reading this, should)

- Mathia